analyzer: knock out real-concern bugprone-* in xtext/hfs/cicn/macres
Step 6 from docs/analyze-triage.md. Filters the four newly-owned
files for the categories the triage classified as real concerns
(implicit-widening-of-multiplication-result, assert-side-effect,
integer-division, branch-clone, signed-char-misuse, reserved-
identifier, not-null-terminated-result, assignment-in-if-condition)
and clears every actionable finding. Noise categories
(narrowing-conversions, casting-through-void, easily-swappable-
parameters, switch-missing-default-case) are intentionally
skipped — they're GObject idioms and stylistic noise.

hfs.c (and hfs.h):

- SIZEOF_HFS_HDR_DESCR retyped as ((size_t)12) so all 18
  multiplication sites compute in size_t. Six read()-return
  comparisons cast the RHS to ssize_t to preserve sign-correctness.
- pathbuf[i] == dir_char comparison casts the char to guint8 (x2).
- suffix(path) extraction split out of the if-condition.
- memcpy(buf, tc, 8) NOLINT'd: tc is an 8-byte type+creator
  binary blob, not a C string.

cicn.c:

- 6 pixel-arithmetic sites get (size_t) on one operand. Width
  and height are bounded <=4096 so no overflow risk; the cast
  is silence-only and improves explicitness.

macres.c:

- _e_int16 / _e_int24 / _e_int32 renamed to mr_extract_int*
  to step out of the implementation-reserved leading-underscore
  namespace; the e_int* macros now call the renamed helpers.

xtext.c:

- charlen macro now casts the g_utf8_skip lookup through guchar
  so the 3 'signed char to int' warnings vanish at every call site.
- Pango glyph cursor x_off / PANGO_SCALE now done in double so
  Cairo gets subpixel-precise positions.
- Scrollbar page_size = alloc_h / fontsize now done in double.
- RENDER_FLUSH branch-clone merged with OR.
- is_del macro arg parenthesised (already done in earlier PR,
  consistent here).
- Two g_assert(ent->str_len < sizeof(...)) sites NOLINT'd —
  bugprone-assert-side-effect false-positives on -> chains
  inside g_assert even when sizeof doesn't evaluate them.

Build clean; unit + proto suite (50 tests) green.

+++ src/hfs.h
@@ -60,7 +60,13 @@ struct hfs_hdr_descr {
 	u_int32_t	length PACKED;	/* The length of the entry */
 };
 
-#define SIZEOF_HFS_HDR_DESCR	12
+/* Cast to size_t so the 18 sites that compute
+ *   SIZEOF_HFS_HDR_DESCR * <int-or-uint16>
+ * do the multiplication in size_t rather than int. Values stay
+ * tiny (12 * HFS_HDR_MAX = 192) so the cast is purely about
+ * silencing bugprone-implicit-widening-of-multiplication-result
+ * — there's no real overflow risk. */
+#define SIZEOF_HFS_HDR_DESCR	((size_t)12)
 
 /* 
  * Default header layout for Netatalk and AppleDouble
